Move remaining stuff to common module
