+++ src/modules/common/actions/login.js
@@ -1,10 +1,11 @@
-import * as types from '../constants/actionTypes';
 import auth from 'panoptes-client/lib/auth';
 import oauth from 'panoptes-client/lib/oauth';
-import config from '../constants/config';
 
-//Action creators
+import * as types from '../../../constants/actionTypes';
+import config from '../../../constants/config';
 
+
+// Action creators
 export function checkLoginUser() {  //First thing on app load - check if the user is logged in.
   return (dispatch) => {
     auth.checkCurrent()
