+++ src/modules/common/containers/MapSelector.jsx
@@ -1,6 +1,6 @@
-const config = require('../constants/mapExplorer.config.json');
+const config = require('../../../constants/mapExplorer.config.json');
 
-export default class MapSelector {
+class MapSelector {
   constructor() {
     this.id = //Random ID.
       '0123456789abcdef'[Math.floor(Math.random() * 16)] +
@@ -12,7 +12,7 @@ export default class MapSelector {
       '0123456789abcdef'[Math.floor(Math.random() * 16)] +
       '0123456789abcdef'[Math.floor(Math.random() * 16)];
     this.mode = MapSelector.GUIDED_MODE;
-    
+
     //Default filter selectors
     this.species = [];  //For a pre-set selection, use ['baboon', 'lion'] or etc.
     this.habitats = [];  //For a pre-set selection, use ['limestone', 'miombo'] or etc.
@@ -20,18 +20,18 @@ export default class MapSelector {
     this.dateStart = '';
     this.dateEnd = '';
     this.user = '';
-    
+
     //Default marker styles.
     this.markerColor = '#ff9900';  //For... consistency, this is coLOR instead of coLOUR.
     this.markerSize = '15';
     this.markerOpacity = '0.8';
-    
+
     //These two are the most important values; they're derived from the selector
     //settings above.
     this.sql = this.calculateSql();
     this.css = this.calculateCss();
   }
-  
+
   calculateSql(sqlQueryTemplate = config.cartodb.sqlQueryCountItems, specificCamera = '') {
     return sqlQueryTemplate
       .replace(/{CAMERAS}/ig, config.cartodb.sqlTableCameras)
@@ -40,10 +40,10 @@ export default class MapSelector {
       .replace(/{AGGREGATIONS}/ig, config.cartodb.sqlTableAggregations)
       .replace(/{WHERE}/ig, this.calculateSqlWhereClause(specificCamera));
   }
-  
+
   calculateSqlWhereClause(specificCamera = '') {
     //The biggest variable in the SQL query is the 'WHERE' clause.
-    
+
     //Where constructor: species
     let sqlWhere_species = [];
     config.species.map((item) => {
@@ -53,8 +53,8 @@ export default class MapSelector {
     });
     sqlWhere_species = (sqlWhere_species.length === 0)
       ? '' : `( ${sqlWhere_species.join(' OR ')} )`;
-    
-    
+
+
     //Where constructor: habitats
     let sqlWhere_habitats = [];
     config.habitats.map((item) => {
@@ -64,7 +64,7 @@ export default class MapSelector {
     });
     sqlWhere_habitats = (sqlWhere_habitats.length === 0)
       ? '' : `(${sqlWhere_habitats.join(' OR ')})`;
-    
+
     //Where constructor: seasons
     let sqlWhere_seasons = [];
     config.seasons.map((item) => {
@@ -74,7 +74,7 @@ export default class MapSelector {
     });
     sqlWhere_seasons = (sqlWhere_seasons.length === 0)
       ? '' : `(${sqlWhere_seasons.join(' OR ')})`;
-    
+
     //Where constructor: date range
     let sqlWhere_dates = '';
     let validDate = /^\s*\d\d\d\d[\-\/\.]?\d?\d[\-\/\.]?\d?\d\s*$/g;
@@ -85,17 +85,17 @@ export default class MapSelector {
     } else if (this.dateEnd.match(validDate)) {
       sqlWhere_dates = '(to_timestamp(dateutc, \'MM/DD/YYYY\') <= \''+this.dateEnd.trim()+'\')';
     }
-    
+
     //Where constructor: user/groups/etc
     let sqlWhere_user = (this.user.trim() === '')
       ? '': '(user_hash ILIKE \''+this.user.replace(/'/, '\'\'').trim()+'\')';
       //WARNING: Either use =, LIKE or ILIKE (case-insensitive Like) depending on what the Panoptes API demands.
-    
+
     //Where constructor: specific camera
     let sqlWhere_camera = (specificCamera.trim() !== '')
       ? '(camera ILIKE \''+specificCamera.replace(/'/, '\'\'').trim()+'\')'
       : '';
-    
+
     //Join the Where constructor
     let sqlWhere = '';
     [sqlWhere_species, sqlWhere_habitats, sqlWhere_seasons, sqlWhere_dates, sqlWhere_user, sqlWhere_camera].map((wherePart) => {
@@ -107,20 +107,20 @@ export default class MapSelector {
     if (sqlWhere !== '') {
       sqlWhere = ` WHERE ${sqlWhere}`;
     }
-    
+
     return sqlWhere;
   }
-  
+
   calculateCss() {
     let children = '[count=0],[count=null] {marker-fill:#000;} ';
-    
+
     //For presentation, TODO: give options to change styles
     //----------------
     for (let i = 0; i < 20; i++) {
       children += `[count>${i*50}] {marker-width:${(parseFloat(this.markerSize)*(1 + i / 10))};} `;
     }
     //----------------
-    
+
     return config.cartodb.cssStandard
       .replace(/{LAYER}/ig, config.cartodb.sqlTableCameras)  //Actually, any ID will do
       .replace(/{MARKERCOLOR}/ig, this.markerColor)
@@ -128,7 +128,7 @@ export default class MapSelector {
       .replace(/{MARKERSIZE}/ig, this.markerSize)
       .replace(/{CHILDREN}/ig, children);
   }
-  
+
   copy() {
     let newCopy = new MapSelector();
     for (let i in this) {
@@ -144,3 +144,5 @@ export default class MapSelector {
 //like checkboxes, etc) or Advanced Mode (which exposes full SQL capabilities).
 MapSelector.GUIDED_MODE = 1;
 MapSelector.ADVANCED_MODE = 2;
+
+export default MapSelector;
