+++ src/modules/teachers/components/ClassroomsOverview.jsx
@@ -1,6 +1,6 @@
 import { PropTypes } from 'react';
 
-import Spinner from '../../../presentational/Spinner.jsx'
+import Spinner from '../../common/components/Spinner.jsx'
 
 
 const ClassroomsOverview = (props, context) => {
