--- src/presentational/Spinner.jsx
@@ -1,7 +1,0 @@
-import { PropTypes } from 'react';
-
-const Spinner = (props) => (
-  <i className="fa fa-cog fa-spin fa-3x"></i>
-);
-
-export { Spinner as default }
